==> code/00_initialize.R <==
library(caret)
library(ebal)
library(raster)
library(SearchTrees)
library(splitstackshape)
library(cowplot)
library(modelsummary)
library(nnet)
library(DescTools)
library(miceadds)
library(readxl)
library(extrafont)
library(stargazer)
library(scales)
library(fixest)
library(rgdal)
library(tigris)
library(Matching)
library(kevostools)
library(rgeos)
library(RSQLite)
library(tidyverse)
library(data.table)

save <- c("db", "cleanup", "theme_bc", "save", "weighted.ttest.ci")
options("modelsummary_format_numeric_latex" = "plain")

cleanup <- function(...){
  save2 <- c(save, ...)
  rm(list=ls(envir = .GlobalEnv)[! ls(envir = .GlobalEnv) %in% save2], envir = .GlobalEnv)
  gc()
}

db <- dbConnect(SQLite(), "E:/national_file.db")
